pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-scheduled-event-pipeline"
version = "0.1.0"
description = "Pipeline + backtest for scheduled-event Polymarket contracts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0",
  "openai>=1.0",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "tqdm>=4.65",
  "numpy>=1.24",
  "httpx>=0.27",
  "polars>=1.0",
  "scipy>=1.11",
]

[project.scripts]
polymarket-backtest = "polymarket_backtest.main:main"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["polymarket_pipeline*", "polymarket_backtest*", "polymarket_backtester*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
